Display foreign key and index labels
To get the foreign keys, we needed a list of relationships from the
table whose node is being built, which required correcting the From/To
confusion in the existing GetRelationshipsFrom method (which was actually
To) and adding a new method.

Also correct the ordering of the columns in the table, which otherwise
ended up with all primary key columns at the beginning, even if that's
not really what their ordinal position was.